+++ .gitignore
@@ -7,7 +7,7 @@
 *.dylib
 
 cmd/tailscale/tailscale
-cmd/tailscaled/tailscaled
+cmd/tailscaled/tailscaled/tailscaled
 
 # Test binary, built with `go test -c`
 *.test
