+++ cmd/tailscaled/tailscaled/tailscaled_windows.go
@@ -2,7 +2,7 @@
 // Use of this source code is governed by a BSD-style
 // license that can be found in the LICENSE file.
 
-package main // import "tailscale.com/cmd/tailscaled"
+package tailscaled
 
 // TODO: check if administrator, like tswin does.
 //
